Replace Java specific API with kotlin multiplat compatible API
This allows importing the code in a separate gradle project, that can
be used for cross-compilation. Current usecase inclue documentation,
and AfterEffects plugin.

- Explicitly import @JvmInline. It is ignored in multiplat code, but
  requires the import
- use `check()` instead of JVM specific `assert()`. The current use of
  `assert` over `check` was not deliberate
- Drop java class references. For SemanticKey, the actual class was
  never used.
- Replace java Object with kotlin Any
- Use `hashCode` instead of System.identityHashCode. Since this is for
  debug messages only, not a big deal.
- Replace MotionValue AtomicInteger use with a plain int. Its used from
  one thread only anyways.
- Use Double.nextUp/nextDown (instead of float)

Bug: 390325138
Test: Existing unit tests
Flag: EXEMPT PURE_REFACTOR

+++ mechanics/src/com/android/mechanics/spec/builder/EffectPlacement.kt
@@ -19,6 +19,7 @@ package com.android.mechanics.spec.builder
 import androidx.compose.ui.util.packFloats
 import androidx.compose.ui.util.unpackFloat1
 import androidx.compose.ui.util.unpackFloat2
+import kotlin.jvm.JvmInline
 import kotlin.math.max
 import kotlin.math.min
 import kotlin.math.nextDown
@@ -72,13 +73,15 @@ value class EffectPlacement internal constructor(val value: Long) {
             }
         }
 
-    internal val sortOrder: Float
+    // TODO Convert this to a Comparable instead. Currently uses Double instead of Float, since
+    //  nextUp / nextDown is only available on Doubles in Kotlin Multiplatform.
+    internal val sortOrder: Double
         get() {
             return when (type) {
-                EffectPlacemenType.At -> start
-                EffectPlacemenType.Before -> start.nextDown()
-                EffectPlacemenType.After -> start.nextUp()
-                EffectPlacemenType.Between -> (start + end) / 2
+                EffectPlacemenType.At -> start.toDouble()
+                EffectPlacemenType.Before -> start.toDouble().nextDown()
+                EffectPlacemenType.After -> start.toDouble().nextUp()
+                EffectPlacemenType.Between -> (start.toDouble() + end.toDouble()) / 2
             }
         }
